+++ src/analysis/templating/topology.rs
@@ -627,6 +627,42 @@ pub fn rank_templates_aux<'t>(mut templates: HashMap<&'t str,
                         debug!("Implicit import of '{}' missing and intentionally ignored",
                                missing_template_name);
                     },
+                    InferiorVariant::InEach(decl) => {
+                        if !BUILTIN_TEMPLATES.iter().any(
+                            |name|name==missing_template_name) {
+                            // The 'is' may contain more names than we
+                            // are looking for, find the particular name
+                            // that matches and use that span
+                            // Because the same name could be used multiple
+                            // times, but this only causes inferior binding
+                            // report one error per such span
+
+                            let spans: Vec<_> = decl.obj.spec.iter()
+                                .filter(|dmlname|
+                                        &dmlname.val == missing_template_name)
+                                .map(|dmlname|dmlname.span())
+                                .collect();
+                            if spans.is_empty() {
+                                internal_error!("Unexpectedly no name \
+                                                 matching missing template \
+                                                 in {:?} (wanted {})",
+                                                decl, missing_template_name);
+                                continue;
+                            }
+                            for span in spans {
+                                report.push(
+                                    DMLError {
+                                        span: *span,
+                                        description: format!(
+                                            "No template; '{}'",
+                                            missing_template_name),
+                                        related: vec![],
+                                        severity: Some(
+                                            DiagnosticSeverity::ERROR),
+                                    });
+                            }
+                        }
+                    },
                     inf => {
                         internal_error!(
                             "Unexpected template dependency through {:?}",